tests: move sharding higher so smokehouse-bundle can also shard
