+++ lighthouse-cli/test/smokehouse/frontends/smokehouse-bin.js
@@ -13,7 +13,6 @@
 
 /* eslint-disable no-console */
 
-import {strict as assert} from 'assert';
 import path from 'path';
 import fs from 'fs';
 import url from 'url';
@@ -23,7 +22,7 @@ import yargs from 'yargs';
 import * as yargsHelpers from 'yargs/helpers';
 import log from 'lighthouse-logger';
 
-import {runSmokehouse} from '../smokehouse.js';
+import {runSmokehouse, getShardedDefinitions} from '../smokehouse.js';
 import {updateTestDefnFormat} from './back-compat-util.js';
 import {LH_ROOT} from '../../../../root.js';
 
@@ -78,52 +77,6 @@ function getDefinitionsToRun(allTestDefns, requestedIds, {invertMatch}) {
   }
 
   return smokes;
-}
-
-/**
- * Parses the cli `shardArg` flag into `shardNumber/shardTotal`. Splits
- * `testDefns` into `shardTotal` shards and returns the `shardNumber`th shard.
- * Shards will differ in size by at most 1.
- * Shard params must be 1 ≤ shardNumber ≤ shardTotal.
- * @param {Array<Smokehouse.TestDfn>} testDefns
- * @param {string=} shardArg
- * @return {Array<Smokehouse.TestDfn>}
- */
-function getShardedDefinitions(testDefns, shardArg) {
-  if (!shardArg) return testDefns;
-
-  // eslint-disable-next-line max-len
-  const errorMessage = `'shard' must be of the form 'n/d' and n and d must be positive integers with 1 ≤ n ≤ d. Got '${shardArg}'`;
-  const match = /^(?<shardNumber>\d+)\/(?<shardTotal>\d+)$/.exec(shardArg);
-  assert(match && match.groups, errorMessage);
-  const shardNumber = Number(match.groups.shardNumber);
-  const shardTotal = Number(match.groups.shardTotal);
-  assert(shardNumber > 0 && Number.isInteger(shardNumber), errorMessage);
-  assert(shardTotal > 0 && Number.isInteger(shardTotal));
-  assert(shardNumber <= shardTotal, errorMessage);
-
-  // Array is sharded with `Math.ceil(length / shardTotal)` shards first
-  // and then the remaining `Math.floor(length / shardTotal) shards.
-  // e.g. `[0, 1, 2, 3]` split into 3 shards is `[[0, 1], [2], [3]]`.
-  const baseSize = Math.floor(testDefns.length / shardTotal);
-  const biggerSize = baseSize + 1;
-  const biggerShardCount = testDefns.length % shardTotal;
-
-  // Since we don't have tests for this file, construct all shards so correct
-  // structure can be asserted.
-  const shards = [];
-  let index = 0;
-  for (let i = 0; i < shardTotal; i++) {
-    const shardSize = i < biggerShardCount ? biggerSize : baseSize;
-    shards.push(testDefns.slice(index, index + shardSize));
-    index += shardSize;
-  }
-  assert.equal(shards.length, shardTotal);
-  assert.deepEqual(shards.flat(), testDefns);
-
-  const shardDefns = shards[shardNumber - 1];
-  console.log(`In this shard (${shardArg}), running: ${shardDefns.map(d => d.id).join(' ')}\n`);
-  return shardDefns;
 }
 
 /**
